Implement getting permission and location

+++ app/src/main/java/com/example/main/domain/usecases/GetWeatherByLocationUseCase.kt
@@ -3,7 +3,7 @@ package com.example.main.domain.usecases
 import com.example.main.domain.models.WeatherInfo
 
 class GetWeatherByLocationUseCase {
-    fun execute(latitude: Int, longitude: Int) : WeatherInfo {
+    fun execute(latitude: Double, longitude: Double) : WeatherInfo {
         return WeatherInfo(cityName = "Kazan", temperatureInCelsius = 0, humidityAsPercentage = 74, pressureInMillimetersOfMercury = 756, windSpeedInMetersPerSecond = 5.9f)
     }
 }
